+++ CHANGELOG.md
@@ -6,12 +6,22 @@ This change log file is based on best practices from [Keep a Changelog](http://k
 This project adheres to [Semantic Versioning](http://semver.org/). Breaking changes result in a different MAJOR version. UI changes that might break customizations on top of the SDK will be treated as breaking changes too.
 This project adheres to the Node [default version scheme](https://docs.npmjs.com/misc/semver).
 
-## [Next-version]
+## [Next-version (Major)]
+
+### Added
+
+- UI: Add country selection screen after document selection. This screen is skipped by default for a preselected document but can still be displayed by enabling the `showCountrySelection` option for the `document` step.
+- Internal: The release script and the `release/RELEASE_GUIDELINE.md` file now include the information needed to update the `MIGRATION.md` file.
+- UI: Added Residence Permit option for document selection
+
+### Changed
+
+- Internal: Changed resolution constraints for live document captures from `720` to `1080`.
 
 ### Fixed
 
+- Public: Return a generic error for unmapped Onfido API validation keys.
 - Fix typo in PhoneNumberInput SASS styles producing invalid CSS
-
 
 ## [5.13.0] - 2020-08-24